Fixed bug where the "Search GitHub" popup always referenced the first tab, regardless of which tab was opened.
Cause: The current tab is cached when the JReaderPanelPopupListener is
created. At the time of creation, the current tab is the only tab.
Later, when creating popup menus, this cached tab is used to get the
filename, which will be incorrect on tabs other than the first.

This change removes the initial caching of the JReaderPanel, and instead
looks up the current tab on demand.